+++ packages/scenes-app/src/demos/dynamicPage.tsx
@@ -14,6 +14,7 @@ import { getQueryRunnerWithRandomWalkQuery } from './utils';
 
 export function getDynamicPageDemo(defaults: SceneAppPageState): SceneAppPage {
   const defaultTabs = [getSceneAppPage('/tab1', 'Temperature')];
+  console.log('get dynamic page demo');
 
   const page = new SceneAppPage({
     ...defaults,
