+++ packages/scenes/src/components/SceneApp/SceneAppPage.tsx
@@ -5,8 +5,8 @@ import { SceneComponentProps } from '../../core/types';
 import { EmbeddedScene } from '../EmbeddedScene';
 import { SceneFlexItem, SceneFlexLayout } from '../layout/SceneFlexLayout';
 import { SceneReactObject } from '../SceneReactObject';
-import { SceneAppPageView } from './SceneAppPageView';
-import { SceneAppDrilldownView, SceneAppPageLike, SceneAppPageState } from './types';
+import { SceneAppDrilldownViewRender, SceneAppPageView } from './SceneAppPageView';
+import { SceneAppPageLike, SceneAppPageState } from './types';
 import { renderSceneComponentWithRouteProps } from './utils';
 
 /**
@@ -114,17 +114,6 @@ function SceneAppPageRenderer({ model, routeProps }: SceneAppPageRendererProps) 
   routes.push(getFallbackRoute(model, routeProps));
 
   return <Switch>{routes}</Switch>;
-}
-
-export interface SceneAppDrilldownViewRenderProps {
-  drilldown: SceneAppDrilldownView;
-  parent: SceneAppPageLike;
-  routeProps: RouteComponentProps;
-}
-
-function SceneAppDrilldownViewRender({ drilldown, parent, routeProps }: SceneAppDrilldownViewRenderProps) {
-  const scene = drilldown.getPage(routeProps.match, parent);
-  return renderSceneComponentWithRouteProps(scene, routeProps);
 }
 
 function getFallbackRoute(page: SceneAppPage, routeProps: RouteComponentProps) {
